+++ tests/integration/test_phase5_dashboard_accessibility_smoke.py
@@ -45,6 +45,9 @@ def test_real_tk_workflow_has_text_and_keyboard_targets_at_common_scaling(
             on_cancel=_callback,
             on_discover=_callback,
             on_export=_callback,
+            on_modify=_callback,
+            on_repeat=_callback,
+            on_new_test=_callback,
             on_close=_callback,
         )
         root.update_idletasks()
@@ -59,6 +62,10 @@ def test_real_tk_workflow_has_text_and_keyboard_targets_at_common_scaling(
             widgets.run_button,
             widgets.discover_button,
             widgets.export_button,
+            widgets.modify_button,
+            widgets.repeat_button,
+            widgets.new_test_button,
+            widgets.finish_button,
             widgets.result_widgets.cancel_button,
             widgets.result_widgets.plot_table,
         )